Move covariate balancing code from `evaluation` to `metrics`
