+++ causallib/evaluation/predictions.py
@@ -9,10 +9,9 @@ import warnings
 
 import pandas as pd
 from ..utils.stat_utils import robust_lookup
-from .metrics import (
-    calculate_covariate_balance,
-    evaluate_metrics,
-)
+from .metrics import evaluate_metrics
+from ..metrics.weight_metrics import calculate_covariate_balance
+
 
 WeightEvaluatorScores = namedtuple(
     "WeightEvaluatorScores", ["prediction_scores", "covariate_balance"]
